// schematics/src/extension/schema.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface PwaExtensionOptionsSchema {
  path?: string;
  tsext?: string;
  project?: string;
  restricted?: boolean;
  /**
   * The name of the extension.
   */
  name?: string;
}